+++ src/Factory/Exception/Client400BadContentExceptionFactory.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace App\Factory\Exception;
 
 use App\Exception\Client400BadContentException;
+use Safe\Exceptions\JsonException;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 
 class Client400BadContentExceptionFactory
@@ -34,6 +35,25 @@ class Client400BadContentExceptionFactory
                 $expectedContent,
                 $realContent
             )
+        );
+    }
+
+    public function createFromJsonException(JsonException $jsonException): Client400BadContentException
+    {
+        return new Client400BadContentException(
+            $this->urlGenerator->generate(
+                'exception-detail',
+                [
+                    'code' => '400',
+                    'name' => 'bad-content',
+                ],
+                UrlGeneratorInterface::ABSOLUTE_URL
+            ),
+            detail: sprintf(
+                'Unable to parse request as JSON. %s.',
+                $jsonException->getMessage()
+            ),
+            previous: $jsonException
         );
     }
 }